# PERF: obj_with_exclusions