--- a/pandas/core/base.py
+++ b/pandas/core/base.py
@@ -221,7 +221,11 @@ class SelectionMixin(Generic[FrameOrSeries]):
             return self.obj[self._selection_list]
 
         if len(self.exclusions) > 0:
-            return self.obj.drop(self.exclusions, axis=1)
+            # equivalent to `self.obj.drop(self.exclusions, axis=1)
+            #  but this avoids consolidating and making a copy
+            return self.obj._drop_axis(
+                self.exclusions, axis=1, consolidate=False, only_slice=True
+            )
         else:
             return self.obj