```
from asv_bench.benchmarks.groupby import *
self = Cumulative()
self.setup("int64", "cumsum")

%timeit self.setup("int64", "cumsum"); self.time_frame_transform("int64", "cumsum")
107 ms ± 7.57 ms per loop (mean ± std. dev. of 7 runs, 10 loops each)  # <- master
74.6 ms ± 2.04 ms per loop (mean ± std. dev. of 7 runs, 10 loops each)  # <- PR
```